Use \`libc\` To Get Passwd/Group on All Unixes
