+++ src/unix/apple/users.rs
@@ -1,72 +1,5 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
-use crate::{
-    common::{Gid, Uid},
-    User, UserInner,
-};
-
-use libc::{c_char, endpwent, getpwent, setpwent, strlen};
-use std::collections::HashMap;
-
-fn endswith(s1: *const c_char, s2: &[u8]) -> bool {
-    if s1.is_null() {
-        return false;
-    }
-    unsafe {
-        let mut len = strlen(s1) as isize - 1;
-        let mut i = s2.len() as isize - 1;
-        while len >= 0 && i >= 0 && *s1.offset(len) == s2[i as usize] as _ {
-            i -= 1;
-            len -= 1;
-        }
-        i == -1
-    }
-}
-
-pub(crate) fn get_users(users: &mut Vec<User>) {
-    fn filter(shell: *const c_char, uid: u32) -> bool {
-        !endswith(shell, b"/false") && !endswith(shell, b"/uucico") && uid < 65536
-    }
-
-    users.clear();
-
-    let mut users_map = HashMap::with_capacity(10);
-
-    unsafe {
-        setpwent();
-        loop {
-            let pw = getpwent();
-            if pw.is_null() {
-                // The call was interrupted by a signal, retrying.
-                if std::io::Error::last_os_error().kind() == std::io::ErrorKind::Interrupted {
-                    continue;
-                }
-                break;
-            }
-
-            if !filter((*pw).pw_shell, (*pw).pw_uid) {
-                // This is not a "real" or "local" user.
-                continue;
-            }
-            if let Some(name) = crate::unix::utils::cstr_to_rust((*pw).pw_name) {
-                if users_map.contains_key(&name) {
-                    continue;
-                }
-
-                let uid = (*pw).pw_uid;
-                let gid = (*pw).pw_gid;
-                users_map.insert(name, (Uid(uid), Gid(gid)));
-            }
-        }
-        endpwent();
-    }
-    for (name, (uid, gid)) in users_map {
-        users.push(User {
-            inner: UserInner::new(uid, gid, name),
-        });
-    }
-}
-
 // This was the OSX-based solution. It provides enough information, but what a mess!
 // pub fn get_users_list() -> Vec<User> {
 //     let mut users = Vec::new();
